moved dailysetup to its own file, and added some volcano quest support
